chore: move CLI to monorepo root
This change moves the CLI Go project to the monorepo and sets a replace
directive to pull in the local gen package in the server. Additionally,
it has been configured with the appropriate build/lint/test mise
tasks in place of the previous tidy/release/version tasks.
