+++ cli/internal/app/build.go
@@ -1,0 +1,6 @@
+package app
+
+var (
+	Version = "dev"
+	GitSHA  = "unset"
+)
