+++ cli/internal/deplconfig/config.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"slices"
 
-	"github.com/speakeasy-api/gram/server/cmd/cli/gram/env"
+	"github.com/speakeasy-api/gram/cli/internal/env"
 )
 
 type SourceType string
