+++ cli/internal/deplconfig/source_reader_test.go
