+++ cli/internal/deplconfig/validate.go
