+++ cli/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "cli",
+  "version": "0.1.3",
+  "private": true
+}
