+++ pnpm-workspace.yaml
@@ -4,6 +4,7 @@ packages:
   - client/*
   - sdks/*
   - server
+  - cli
 
 catalog:
   '@ai-sdk/openai': ^1.3.4
