--- server/cmd/cli/gram/version/version.go
@@ -1,3 +1,0 @@
-package version
-
-const Version = "v0.1.3"
